change to double add infinity

+++ include/graphics.h
@@ -19,7 +19,7 @@ void		setup_window(t_map *map);
 uint32_t	write_color(double r, double g, double b, double a);
 uint32_t	vec_to_color(t_vector vec);
 
-int			hit_sphere(t_object sp, t_ray ray, int *pos, int *t);
+int			hit_sphere(t_object sp, t_ray ray, int *pos, double *t);
 void		hit(t_map *map, int j, int i);
 
 t_ray		get_ray(t_camera *camera, double x, double y);
